Made changes to HsDecls to support multiple equations in a type family group. Now there is a TyFamInstEqn type that tracks the details of a single family instance; the TyFamInstDecl type contains a list of TyFamInstEqn's. To make this change, I also separated data instances from type instances, making a new DataFamInstDecl type. Another knock-on effect (that Simon PJ approved) was to specialize HsTyDefn (which used to track both 'type' and 'data' declarations) just to data declarations.
A type family group with multiple instance equations is rejected
by the type-checker, as Core has no way to deal with such a thing
(yet). The pretty-printer can deal with multiple equations (printing
them under a 'type instance where' header) as can the renamer.
Template Haskell cannot deal with them (yet).

One bugfix: It used to be that you could make an associated
default data family instance using Template Haskell. This
caused a panic (because associated default data family instances
are not allowed). (Normally, an attempt to make such aa thing
is not allowed to parse.) This bug is fixed, and Convert.lhs
reports an error if you try this.