Added overlapping type family instances.
The primary change is in compiler/types/FamInstEnv.hs, where the new
lookup algorithm is written. See comments in that file for the full
details. The biggest change that could affect other people is that,
now, the top-level family instance declaration leads to a FamInstGroup,
not just a FamInst. A FamInstGroup, in turn, holds an ordered list
of FamInsts. This change required a ripple of changes throughout GHC,
including in interface files, to support it.

Still to do: clean up some code/names, allow code to compile when
it relies on a certain corner-case overlap, emit warnings/errors
when certain branches of an instance group are unreachable.