Added check to find inaccessible equations in family instance groups.
This check uses the isDominatedBy function in types/FamInstEnv, which
is currently incomplete. Thus, it is possible some inaccessible
equations will not be marked as such. However, all equations marked
as inaccessible are, so good code can never produce an error.
It's worth noting that bad code that does not produce an error is
silly, but does not compromise the type system.