Removed confluent overlap check from branched type family instances.
It was discovered that confluent overlap within groups implemented
the "obvious" way doesn't support substitution. We (Simon PJ,
Stephanie Weirich, Conor McBride, and I) were unable to come up with
a simple solution to this problem. In any case, we'll wait until
someone shouts loudly enough to make this worthwhile.