Checkpoint on changing over Type and Coercion representation to "nokinds".
Most of the changes are changing function names from ...TyVar...
to ...TyCoVar... This change is being done to make it explicit when
we are dealing with all compile-time variables, or just type variables.

Also of note is that TypeRep has been renamed TyCoRep, and the
Coercion representation (and substitution functions) is there too.

Also of note is that TvSubst has become TCvSubst; CvSubst is done away
with.

These changes are all necessary because types may now contain coercion
variables.