Extended pattern-matches on Type, by searching for ForAllTy.
This commit includes the extension of normaliseType to the new forms
of Type, requiring changes to the code in liftCoSubstVarBndr. See
Note [Normalising types] in FamInstEnv for details.

Panics on seeing new type forms in:
  tcLiftCoSubstWith
  normaliseFfiType'
  flatten

Also, canEq probably needs to deal with the new type forms, but it
won't panic.

vectType uses pprSorry to deal with the new type forms.

Still waiting to talk to GHC HQ about isTauTy and functorLikeTraverse.