Added trace statements to help understand T7368.
The problem with T7368 is that the typechecker will create
*heterogeneous* equalities as it tries to simplify constraints.
Then, it tries to use these equalities to perform substitution,
but that's very silly. Proposal to fix: perform the substitution,
but cast the resultant type back to have the right kind. This is
tricky, though, because the (not yet derived) coercion is a *lifted*
equality. So, it's not obvious how to have a type depend on a
lifted equality in some way. The other possibility is just not
to perform the substitution at all. In any case, T7368 is an
example of code that shouldn't compile, so for now, there is no
evidence of correct code tickling this problem.