x86_64: Setup frame pointer for ccalls
rbp is usually the frame pointer base register in the C calling
convention. It is typically saved by the caller, so we don't strictly
need to preserve it. However, DWARF stack unwinders typically set $rsp =
$rbp to unwind the stack. If $rbp if left pointing to our StgRegs (as it
typically does in Haskell code), the unwinder will turn overwrite $rsp.
This is problematic as we will need $rsp to continue unwinding into the
C stack after we have reached the top of the STG stack.

For this reason, we need to preserve the value of $rbp and set it to the
current $sp before passing execution to the C target.