Fix grouping for pattern synonyms
When grouping pattern synonyms in the desugarer, to find when a single
match will work for the whole group, we use Match.sameGroup.  But this
function was declaring two pattern-synonym matches equal to often.
Result: Lint errors and broken semantics.

The fix is easy.  See Note [Pattern synonym groups].

Re-do typechecking for pattern synonym signatures