RTS linker refactoring
Summary:
This diff started as a simple fix for a linker issue (#9314) and ended
up with this quite big refactoring to make it easier to ensure that the
fix was correct in every case.

To fix the memory consumption issue in the RTS linker (#9314), we fill
the mmap'ed pages as much as possible instead of allocating one page per
object code section. We obtain huge memory gains (153MB -> 86KB in the
test case given in trac #9314). [Startup performance may be better too
because of the reduced number of mmap calls (~20k -> 470)]

During the refactoring, I figured out that section alignment constraints
were not always correctly handled (iirc, the alignment had to be correct
in the object code file in the first place, which is not always the
case). Now it should always be correctly handled when we load a section.

Test Plan: We should test on Windows, OS X, PowerPC and ARM in addition to Linux/x86-64.

Reviewers: erikd, simonmar, austin, bgamari

Reviewed By: erikd

Subscribers: trofi, kgardas, simonmar, thomie

Differential Revision: https://phabricator.haskell.org/D1470

GHC Trac Issues: #9314