Simplify: Make generated names more useful
makeTrivial is responsible for concocting names during simplification.
Previously, however, it would make no attempt to generate a name
that might be useful to later readers of the resulting Core. Here we add
a bit of state to SimplEnv: a name for the context which we are
currently simplifying within. We then derive generated binders from this
name.

See #11676.