base: Eliminate the realToInteger madness
CTime, CClock, and friends don't have Integral instances under the
Haskell 2010 Report. Unfortunately, this means that we had to jump
through all kinds of crazy hoops to get an Integer out of these types.

Here I rework these madness to simply peel away the newtype wrapper and
assume that we have something Integral inside (which we really must, as
these are morally integers).