OccName: Encode isDerived explicitly
Previously we would distinguish the "derived-ness" of an OccName by
looking at the name itself. However, this requires unpacking the name,
which isn't terribly cheap. Instead we now encode the "derived-ness" in
a dedicated flag carried by OccName.

In order to avoid increasing the size of the OccName constructor I
encode it packed alongside the namespace.