Handle deserialization of tuples, etc specifically
Previously BinIface would build a UniqFM containing all of the known-key
things and use this to resolve the Names of known-key things during
interface file deserialization. However, this has the disadvantage of
building up a rather large map containing all type and data constructors
for anonymous sums and tuples.

We now instead add codepaths to map a unique back to its associated
known key name, allowing us to greatly shrink the size of the map.