build system: Persist CrossCompiling in binary distributions
The build system uses the CrossCompiling variable to decide whether or not we
should build various packages that must be built using the compiler.
Consequently, it is important that we persist its value in the binary
distribution so we know during `make install` not to go looking for files that
would have been built for these packages. Failing to do this causes #13325.