StgLint: Always add case binder to in-scope set
Previously we wouldn't add the binder to the in-scope set if linting a
MultiValAlt-flavoured case. However, this neglects to consider that the
scrutinee may not be a unboxed tuple or sum but instead a void argument
(e.g.  arity-0). In this case it is safe to use case binder (and indeed
we do so).

Reviewers: austin

Subscribers: simonpj, rwbarton, thomie

GHC Trac Issues: #14118

Differential Revision: https://phabricator.haskell.org/D3858