StgLint: Enforce MultiValAlt liveness invariant only after unariser
The unariser ensures that we never use case binders that are void, unboxed sums,
or unboxed tuples. However, previously StgLint was enforcing this invariant even
before the unariser was running, giving rise to spurious lint failures. Fix
this. Following CoreLint, we introduce a LintFlags environment to the linter
monad, allowing for additional flags to be easily accomodated in the future.

See #14118.