Model divergence of retry# as ThrowsExn, not Diverges
The demand signature of the retry# primop previously had a Diverges result.
However, this caused the demand analyser to conclude that a program of the
shape,

    catchRetry# (... >> retry#)

would diverge. Of course, this is plainly wrong; catchRetry#'s sole reason to
exist is to "catch" the "exception" thrown by retry#. While catchRetry#'s demand
signature correctly had the ExnStr flag set on its first argument, indicating
that it should catch divergence, the logic associated with this flag doesn't
apply to Diverges results. This resulted in #14171.

The solution here is to treat the divergence of retry# as an exception. Namely,
give it a result type of ThrowsExn rather than Diverges.