configure: Fix ar probed flags
FP_PROG_AR_ARGS appears to be probing for what flags to use to build an archive
with an index. However, none of the flags that it probed previously were valid
as they all started with `c`, which isn't a mode character. This went unnoticed
until someone attempted to use LLVM ar since we have a special case for gnu ar.

Additionally, we work around a bug in llvm-ar 5.0 where ar will exit with code 0
even if it fails to parse the command line. Silliness.

I believe these should rather all be mode `q` however I'll need to test this on
a few platforms to be certain.