Remove special case for WEAK in nonmoving_scavenge
By the time we scavenge a WEAK the key may already dead (e.g. died in a
minor collection) so there may be nothing to evacaute.

Really the special case was just wrong. We already evacuate the whole
weak if key is in non-moving heap, and if it's not then `isAlive`
returns the correct answer, and scavenge does the right thing.

So by the time we mark WEAKs (in `nonmoving_mark_weaks`) if we find that
the key is alive, then we know value is also in the non-moving heap
because we already conservatively return `true` for `isAlive(key)` when
key is in non-moving heap.