Mark large/pinned objects using their bdescr
A large block usually just contains one object so it doesn't matter
whether we mark the bdescr or the object.

However, pinned blocks usually contain multiple objects and they're also
large. In this case to keep the whole block alive during
`nonmoving_sweep_large_objects` we have to mark the bdescr and not
objects.

This is very subtle and should be documented somewhere.