Ensure that we don't loop on heap cycles
One consequence of the new mark criterion is that we need to mark blocks which
have already been marked. This means that we will fail to terminate when faced
with a heap with cyclic closures. Specifically, in the case of closure in an
active segment it's impossible to tell whether it has already been traced
or whether it is merely marked from the last major GC. This is, of course, quite
bad.

Here we add an additional bit of information for each closure which encodes
whether we have marked the closure during the current major GC. To avoid having
to reset this on every GC cycle we use the usual trick of flipping the polarity
of the flag (we call the current polarity the current "mark epoch").