[WIP] CoreToStg: Try treating String unpackings as single-entry
Summary:
This is a quick attempt of the idea proposed in #15113, making top-level
bindings whose RHS is an application of a string unpack operation (e.g.
`unpackCString#`) single-entry. While this reduces sharing of the unpacked
`[Char]`s, it significantly reduces the number of CAFfy bindings in a typical
program (see #15038).

Reviewers: alpmestan

Subscribers: alpmestan, simonpj, rwbarton, thomie, carter

GHC Trac Issues: #15113

Differential Revision: https://phabricator.haskell.org/D4717