Implement "array list" for pointer arrays in the non-moving heap
See #77 for more details. In summary, to be able to do minor GC
concurrently with a non-moving collection, we need a list that has all
arrays in the non-moving. This list should be scavenged in minor GCs. We
remove dead entries in the list in the second and final pause, after
mark phase finishes.