XXX: Hack around another terrible shutdown deadlock
Essentially another version of what we saw previously:

shutdownCapabilities takes all capabilities during shutdown while we are trying
to synchronize the nonmoving collection. This time I tried teaching
shutdownCapability to check the capability's return_tasks list before shutdown
but sadly that also races since we may not have started the synchronization by
the time we start shutting down capabilities.

For now I've just started tightened the race window a bit by checking the
scheduler state before synchronizing. This is terrible and does not fix the
underlying race, but in practice it's much better than it was before. I'll need
to return to this.