Don't create a new HashTable with every upd_rem_set flush
Previously we would call init_mark_queue each time we flushed the update
remembered set. This would result in a new HashTable being allocated, which
would then manifest as the C heap growing to fill the entirely < 4GB address
space, resulting in mmap failures when building aeson (#116).

The right solution to this would be to refactor MarkQueue/UpdRemSet to remove
the unneeded HashTable from the latter (#117). However, in the meantime this
workaround works.