Rely only on liveness bitmap to determine reachability of snapshotted closures
We can determine the reachability of objects which were in the nonmoving heap at
the time that the snapshot was taken more precisely than those which we are
unsure of. The reason is that the snapshot invariant guarantees that such
objects will have their mark bit set at the end of marking.

In this case of TSOs this extra precision matters since we rely on knowing the
reachability of the thread during resurrection. Since we take a snapshot of the
thread list at the beginning of collection (moving oldest_gen->threads to
oldest_gen->old_threads), we know that all threads are in the snapshot.