rts: Add read barrier in deRefWeakPtr#
In general the major GC assumes that all pointers reachable from a
marked object are themselves marked (or in the mark queue). However,
weak pointers are an obvious exception to this rule. In particular,
deRefWeakPtr# allows the mutator to turn a weak reference into a strong
reference. This interacts badly with concurrent collection. For
instance, consider this program:

```
f :: a -> b -> IO b
f k v = do
    -- assume that k and v are the only references to the
    -- closures to which they refer.
    weak <- mkWeakPtr k v Nothing

    -- N.B. k is now technically dead, but we've not yet had a
    -- chance to tombstone the WeakPtr.
    -- This will happen during the major GC.
    performMajorGC

    -- Now we running concurrently with the mark
    Just x <- deRefWeak weak
    -- We have now introduced a reference to `v`, which will
    -- not be marked as the only reference to `v` at t_0 was
    -- via a WeakPtr.
    return x
```