Use https links in user-facing startup and error messages
Summary:
I consider myself lucky that in my circle of friends, `http` urls (as opposed to `https` urls) are frowned upon in that we generally apologize in the rase cases that we share an `http` url.

This pull request changes `http` links into their `https` analogues in the following places:

* In the GHCI startup message (and parts of the User's Guide, where there are verbatim transcripts of GHCi sessions).
* In a couple of error messages, asking the user to report a bug.

(I also took the liberty to change a single space before the reportabug url into two spaces, harmonizing this occurence with the others.)

I'm not trying to start a war. I just had a moment to spare and felt like preparing this diff. Merge or don't merge as you wish!

Reviewers: bgamari, erikd, simonmar

Subscribers: goldfire, rwbarton, carter

Differential Revision: https://phabricator.haskell.org/D5450