Hadrian: introduce userDefaultFlavour, making default flavour overridable
Summary:
This patch introduces the `userDefaultFlavour` user setting. It should
be the name of the default flavour to use when no --flavour argument is
passed. Before this patch, we would just always default to... the `default`
flavour. With this patch, we default to whatever Flavour whose name is
`userDefaultFlavour`, therefore providing a way for users to "persist" their
choice of flavour, not having to repeat --flavour=[...] in every hadrian
command.

Test Plan:
Set `userDefaultFlavour = "quickest"`, run `hadrian/build.sh`, check
that the quickest flavour is indeed picked.

Reviewers: snowleopard, bgamari

Reviewed By: snowleopard

Subscribers: mpickering, rwbarton, carter

GHC Trac Issues: #15890

Differential Revision: https://phabricator.haskell.org/D5454