rts: Specialize hashing at call site rather than in struct.
Summary:
Separate word and string hash tables on the type level, and do not store the hashing function.
Thus when a different hash function is desire it is provided upon accessing the table. This is worst case the same as
before the change, and in the majority of cases is better. Also mark the functions for aggressive inlining to improve performance.
{F1686506}

Reviewers: bgamari, erikd, simonmar

Subscribers: rwbarton, thomie, carter

GHC Trac Issues: #13165

Differential Revision: https://phabricator.haskell.org/D4889