testsuite: Don't disable LLVM ways due to unversioned llc
Summary:
Previously we would silently disable the LLVM testsuite ways when LLC is set to
an unversioned llc executable (e.g. `llc` instead of `llc-5.0`). Presumably
this was because we were lax in checking the LLVM version in configure.
However, this is no longer the case and consequently this confusing behavior is
no longer necessary.

Test Plan: Validate

Reviewers: angerman

Reviewed By: angerman

Subscribers: michalt, rwbarton, thomie, carter

Differential Revision: https://phabricator.haskell.org/D4333