Remove ExnStr and ThrowsExn business
In #14998, Item 1, we noticed that we could get rid of the last usage
of `catchArgDmd` and did so in commit 00b8ecb.

Since then, `catchArgDmd` has been dangling around in Demand.hs. This
commit finally removes it and the resulting dead code involving `ExnStr`
and `ThrowsExn`, which was the original concern of Item 1.